heatmap setting in same path as heatmap view
